dev-requirements.txt
flake8
pytest
pyyaml
requests-mock
yq
